requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
slack-bolt==1.18.1
anthropic==0.28.0
openai==1.35.0
voyageai==0.2.3
pinecone==4.1.0
supabase==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
statsmodels==0.14.6
